breaks useProposal to serpeate hooks and files - moves types and utils to own files - removes conditional preventing loading of proposals
