+++ src/providers/govenor/types/index.ts
@@ -1,3 +1,4 @@
 export * from './govenor';
+export * from './proposal';
 export * from './contractMatch';
 export * from './token';
